+++ packages/ui/src/components/leaflet-map.tsx
@@ -1,0 +1,57 @@
+'use client';
+
+import 'leaflet/dist/leaflet.css';
+
+import L from 'leaflet';
+import { useEffect } from 'react';
+import { MapContainer, Marker, Popup, TileLayer } from 'react-leaflet';
+
+import { cn } from '../lib/utils';
+
+// Fix for default marker icon in Next.js/webpack
+const defaultIcon = L.icon({
+  iconUrl: 'https://unpkg.com/leaflet@1.9.4/dist/images/marker-icon.png',
+  iconRetinaUrl: 'https://unpkg.com/leaflet@1.9.4/dist/images/marker-icon-2x.png',
+  shadowUrl: 'https://unpkg.com/leaflet@1.9.4/dist/images/marker-shadow.png',
+  iconSize: [25, 41],
+  iconAnchor: [12, 41],
+  popupAnchor: [1, -34],
+  shadowSize: [41, 41],
+});
+
+export interface LeafletMapProps {
+  latitude: number;
+  longitude: number;
+  zoom?: number;
+  markerLabel?: string;
+  className?: string;
+}
+
+export function LeafletMap({
+  latitude,
+  longitude,
+  zoom = 15,
+  markerLabel,
+  className,
+}: LeafletMapProps) {
+  useEffect(() => {
+    L.Marker.prototype.options.icon = defaultIcon;
+  }, []);
+
+  return (
+    <MapContainer
+      center={[latitude, longitude]}
+      zoom={zoom}
+      scrollWheelZoom={false}
+      className={cn('h-[300px] w-full rounded-lg', className)}
+    >
+      <TileLayer
+        attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+        url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+      />
+      <Marker position={[latitude, longitude]}>
+        {markerLabel && <Popup>{markerLabel}</Popup>}
+      </Marker>
+    </MapContainer>
+  );
+}